Fixed additional textures not showing up.

+++ assets/shaders/2d_renderer_basic.fs
@@ -3,7 +3,7 @@ out vec4 out_color;
 
 in vec2 uv;
 in vec4 color;
-flat in int texture_unit;
+flat in float texture_unit;
 
 uniform sampler2D texture_samplers[32];
 
@@ -11,6 +11,6 @@ void main() {
     if (texture_unit < 0) {
         out_color = color;
     } else {
-        out_color = texture(texture_samplers[texture_unit], uv) * color;
+        out_color = texture(texture_samplers[int(texture_unit)], uv) * color;
     }
 }
